Move methods into IpFieldMapper
